requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Cors==4.0.0
yfinance==0.2.43
pandas==2.2.2
python-dotenv==1.0.1
yahooquery
requests
